ajustes tipo de key

+++ Escon.SisctNET.Repository/Implementation/ClientRepository.cs
@@ -45,11 +45,11 @@ namespace Escon.SisctNET.Repository.Implementation
             List<string> result = null;
             if (type == "all")
             {
-                result = _context.Clients.Where(_ => _.CompanyId.Equals(companyId) && _.TypeClientId.Equals(1)).Select(_ => _.Document).ToList();
+                result = _context.Clients.Where(_ => _.CompanyId.Equals(companyId) && _.TypeClientId.Equals((long)1)).Select(_ => _.Document).ToList();
             }
             else if (type == "raiz")
             {
-                result = _context.Clients.Where(_ => _.CompanyId.Equals(companyId) && _.TypeClientId.Equals(1)).Select(_ => _.CnpjRaiz).Distinct().ToList();
+                result = _context.Clients.Where(_ => _.CompanyId.Equals(companyId) && _.TypeClientId.Equals((long)1)).Select(_ => _.CnpjRaiz).Distinct().ToList();
             }
             
             AddLog(log);
